requirements.txt
os
glob
numpy==1.23.1
pandas
netCDF4==1.5.8
xarray==2022.3.0
python-dateutil==2.8.2
statsmodels==0.13.2
scikits.statsmodels==0.2.0
scipy==1.8.1
matplotlib==3.5.1
seaborn==0.11.2